requirements.txt
Django==1.6.5
django-haystack==2.0.0
pyelasticsearch==0.6.1
requests==2.2.1
flake8==2.2.0
pep8==1.5.7
